Refactor the entire backend to a modular, vertical-slice architecture.
This change introduces a new directory structure following standard Go project layout conventions (`cmd`, `internal`).

- Each business domain (`auth`, `message`) is now a self-contained module under `internal/`, with its own domain, application, persistence, and presentation layers.
- A `shared` directory under `internal/` now holds all common code, including database connections, the HTTP server, and shared domain models.
- Each module has a `module.go` file that acts as a dependency injection container, responsible for wiring up the module's components and registering its routes.
- The main entry point at `cmd/server/main.go` has been simplified to initialize shared components and then delegate route registration to each module.
- The `Store` interface and repository interfaces have been restructured to the domain layer to resolve circular dependencies.

This new structure significantly improves maintainability, scalability, and separation of concerns within the backend.
