--- backend/application/repositories/message_repository.go
@@ -1,12 +1,0 @@
-package repositories
-
-import (
-	"backend/domain"
-	"context"
-)
-
-type MessageRepository interface {
-	Create(ctx context.Context, message *domain.Message) (*domain.Message, error)
-	FindByID(ctx context.Context, id int) (*domain.Message, error)
-	FindByChannelID(ctx context.Context, channelID int) ([]*domain.Message, error)
-}
