+++ backend/internal/message/module.go
@@ -1,0 +1,31 @@
+package message
+
+import (
+	"backend/internal/message/application"
+	"backend/internal/message/presentation"
+	"backend/internal/shared/infra/db"
+
+	"github.com/gin-gonic/gin"
+	"github.com/go-playground/validator/v10"
+)
+
+// RegisterModule initializes the dependencies for the message module and registers its routes.
+func RegisterModule(
+	store *db.SQLStore,
+	router *gin.RouterGroup,
+	validate *validator.Validate,
+	authMiddleware gin.HandlerFunc,
+) {
+	// Initialize use cases
+	createMessageUseCase := application.NewCreateMessageUseCase(store)
+
+	// Initialize the controller
+	messageController := presentation.NewMessageController(createMessageUseCase, validate)
+
+	// Register routes under an authenticated group
+	messageRoutes := router.Group("/messages")
+	messageRoutes.Use(authMiddleware) // Apply JWT middleware to all message routes
+	{
+		messageRoutes.POST("/", messageController.CreateMessage)
+	}
+}
