+++ backend/internal/shared/domain/errors.go
